refactor: resolve Sonar issue - remove duplication
